Move Navigator to production in Domain module
This was a close call because of `NavigationTopic`'s Codable conformance which kind of leaks SwiftUI details into it. But it doesn't technically depend on SwiftUI. Also, when I had the Navigator and its tests in the main module, I found myself importing Domain everywhere, while I don't have to import the main module now that we're in the Domain module. So it seems more natural to have the Navigator in the Domain module. Finally, the Navigator tests have no need for the iOS simulator to run. They can run on macOS, which is faster. And apart from the Codable conformance, a tiny detail, the Navigator is pretty generic and abstract. So I went with the Domain module for now.