Hide topic store behind protocol for simpler testing
This is like an example in the book "Dependency Inversion Principles, Practices, and Patterns". Normally, a protocol should be in the same module as its consumer, which is the `AppModel` (in the Presentation module) in this case, so that the dependency is inverted. But of course, we don't want the Domain module to depend on the Presentation module. The purpose of the protocol here is not to invert the dependency, but to enable more granular unit testing. With the protocol in place, we can now replace it with a test double when testing the `AppModel`.