When navigating to topic with empty name, show settings and and focus name text field
As long as the name is empty, this will also happen when opening the settings manually or when navigating to an existing (= not new) topic with an empty name. So you could also delete the name of an old topic, and it will show this behavior. This was the simplest way to implement this. It might make sense to keep it this way.